update light and shape

+++ src/base/lightsampler.h
@@ -27,7 +27,7 @@ protected:
 public:
     explicit LightSampler(const LightSamplerDesc &desc);
     [[nodiscard]] span<const Light *const> lights() const noexcept { return _lights; }
-    [[nodiscard]] size_t light_num() const noexcept { return _lights.size(); }
+    [[nodiscard]] uint light_num() const noexcept { return _lights.size(); }
     void add_light(Light *light) noexcept { _lights.push_back(light); }
     [[nodiscard]] virtual Float PMF(const Uint &id) const noexcept = 0;
     [[nodiscard]] virtual SampledLight sample(const Float &u) const noexcept = 0;
